Port ChaosPlan to yaml
Who would have know that our little abstraction would be paying
dividents so soon, with the format of test plan we want the outcome file
is a cluster fuck of json so we have had to port it to yaml in order to
make it human friendly.

docker-compose is yaml already so it is not a big deal.