Remove docker-compose logging
The current logging framework works by running the 'docker-compose logs'
process, reading the output, splitting it based on the container name
and saving it in a different file per container.

This does not work for us as when we scale a container or if the
container dies the running instance of 'docker-compose logs' quits. It
also does not monitor for newly spawned containers.

This new class DockerLogger does not actually log anything for us, but
it will in the future.